+++ HISTORY.rst
@@ -2,7 +2,7 @@
 History
 =======
 
-0.2.5 (YYYY-MM-DD)
+0.2.5 (2020-05-11)
 ------------------
 * Remove deadlock in TableProxy weakref.finalize on Python 3.6 (:pr:`113`)
 * Use python-casacore wheels for travis testing, instead of kernsuite packages (:pr:`115`)